requirements.txt
streamlit
torch
torchvision
timm
numpy
opencv-python
matplotlib
Pillow
scikit-learn
